fix itemLink для предметов экипировки: убрать .int() из валидации itemId
ID предметов экипировки = Date.now()+Math.random() (float),
а zod схема требовала z.number().int() — валидация молча падала.
У craft-предметов ID целые, так что .positive() достаточно.

+++ server/src/validation.ts
@@ -105,7 +105,7 @@ export const wsPrivateMessageSchema = z.object({
 
 export const wsItemLinkSchema = z.object({
   type: z.literal('itemLink'),
-  itemId: z.number().int().positive().optional(),
+  itemId: z.number().positive().optional(),
   itemData: z.any().optional(),
 });
